Split large client and tests into partial classes
